Fix tests & move them to corresponding directory
